v1 code: 1.optimized the playing timing of the steel ruler, improving the playing speed.​ 2.update ruler's F-L formula parameter

+++ v1/code/main/digital_mic.h
@@ -4,7 +4,7 @@
 #include <stdint.h>
 
 extern void i2s_driver_init(void);
-extern int sound_fft_init(float fft_size);
+extern int sound_fft_init(int fft_size);
 extern int get_sound_frequency(float min, float max, float *freq, bool print);
 
 #endif // __DIGITAL_MIC_H
